+++ NEWS.md
@@ -1,5 +1,8 @@
 # dplyr 0.5.0.9000
 
+* `bind_cols()` now calls `tibble::repair_names()` to ensure that all
+  names are unique (#2248).
+
 * You can now recode a factor to other types (#2268)
 
 * `distinct()` no longer duplicates variables (#2001)